+++ src/data/store/arena/bytes/impls/mod.rs
@@ -1,4 +1,4 @@
 // devela/src/data/store/arena/bytes/impls/mod.rs
 
 mod array; // __arena_bytes_impl_array!
-// mod vec; // __arena_bytes_impl_vec!
+mod vec; // __arena_bytes_impl_vec!